% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_time.R
\name{dec_time}
\alias{dec_time}
\title{Calculate time in hh:mm format from decimal time}
\usage{
dec_time(dectime)
}
\arguments{
\item{dectime}{time vector in decimal numeric format}
}
\value{
character time vector in hh:mm format
}
\description{
calculates time in format 23:30 from 23.5
}
\examples{
dt <- c(23.1, 23.5, 23.6)
dec_time(dt)

}
\author{
Maite Lezama Valdes
}
