% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmvt.R
\name{lmvt}
\alias{lmvt}
\title{Calculate UTC and local time from Local Solar Time}
\usage{
lmvt(fnam, viewtime, ra = NULL, aoiproj = NULL, ltz = NULL)
}
\arguments{
\item{ra}{optional: extent of the area of interest, projected in the local projection to be provided in
aoiproj}

\item{aoiproj}{optional, but required if ra is being used: local projection of the area of interest}

\item{ltz}{local time zone,  e.g. "Pacific/Auckland" or "EST" to be found in R function OlsonNames()}

\item{LocST}{Local solar time vector in mm:hh format or as a viewtime raster
in decimal format or decimal format only}

\item{utc}{UTC day that was provided e.g. via filename or the corresponding
UTC day (infered based on the local time zone the
local solar time was reported for conversion to UTC time), format ='\%Y-\%m-\%d'
or as POSIXlt date}

\item{lon}{required only if no raster is used for LocST, to provide as decimal
point numeric}
}
\value{
data frame with lon (and lat in case of raster input), as well as decimal format
local solar time ("LocST"), hour format local solar time "LocST_h", UTC date calculated
from LocST "utc_from_LocST", the original UTC input provided ("orgUTC") and a time
difference between provided and calculated UTC times in case a POSIXlt format UTC
date with time input was provided (available e.g. in case of MODIS L2/swath data)
}
\description{
calculate UTC and local time from Local Solar Time, as provided
e.g. in MODIS viewtime rasters
}
\examples{
library(raster)
vt_raster <-  raster(system.file("extdata", "vtr_s.tif", package="LocST"))
aoi_ext <- extent(-120775.7, -37096.71, 6686761, 6782394)
aoiproj <- "+proj=utm +zone=54 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs"

lmvt(fnam="MOD11_L2.A2018019.1350.006.2018020082352.hdf",
     viewtime = vt_raster, ra=NULL, aoiproj=NULL, ltz="Australia/ACT")

# with Swath data + optionally: cut to area of interest and calculate also local timezone
lmvt(fnam="MOD11_L2.A2018019.1350.006.2018020082352.hdf",
     viewtime = vt_raster, ra=aoi_ext, aoiproj=aoiproj, ltz="Australia/ACT")

lmvt(fnam="MOD11A1.A2018019.h29v12.006.2018020085050.hdf",
     viewtime = vt_raster, ra=NULL, aoiproj=NULL, ltz="Australia/ACT")

}
\author{
Maite Lezama Valdes
}
